Don't create RethinkDB instance if it is already configured in this VM.

+++ Vagrantfile
@@ -24,9 +24,12 @@ Vagrant::Config.run do |config|
         apt-get update --assume-yes;
         apt-get install --assume-yes rethinkdb;
 
-        sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
-                                echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
-        rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+        if [ ! -f /etc/rethinkdb/instances.d/default.conf ];
+        then
+          sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
+          echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
+          rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+        fi
 
         service rethinkdb start;
       EOF
@@ -34,7 +37,7 @@ Vagrant::Config.run do |config|
   end
 
 
-  (1..11).each do |i|
+  (1..3).each do |i|
     config.vm.define "slave#{i}" do |slave_config|
       # Ubuntu 12.04, 64 bit
       slave_config.vm.box     = "precise64"
@@ -47,9 +50,9 @@ Vagrant::Config.run do |config|
 
       slave_config.ssh.username = "vagrant"
       slave_config.vm.network :hostonly, "10.10.10.1#{i}"
-      slave_config.vm.forward_port 8080, 8081
-      slave_config.vm.forward_port 28015, 28016
-      slave_config.vm.forward_port 29015, 29016
+      slave_config.vm.forward_port 8080, (8080 + i)
+      slave_config.vm.forward_port 28015, (28015 + i)
+      slave_config.vm.forward_port 29015, (29015 + i)
 
       slave_config.vm.provision :shell do |sh|
         sh.inline = <<-EOF
@@ -60,10 +63,13 @@ Vagrant::Config.run do |config|
           apt-get update --assume-yes;
           apt-get install --assume-yes rethinkdb;
 
-          sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
-          echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
-          echo "join=10.10.10.10:29015" >> /etc/rethinkdb/instances.d/default.conf
-          rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+          if [ ! -f /etc/rethinkdb/instances.d/default.conf ];
+          then
+            sed -e 's/somebody/root/g' -e 's/somegroup/root/g' /etc/rethinkdb/default.conf.sample > /etc/rethinkdb/instances.d/default.conf
+            echo "bind=all" >> /etc/rethinkdb/instances.d/default.conf
+            echo "join=10.10.10.10:29015" >> /etc/rethinkdb/instances.d/default.conf
+            rethinkdb create -d /var/lib/rethinkdb/instances.d/default 2>&1;
+          fi
 
           service rethinkdb start;
         EOF
